+++ plugins/saihubot-search.js
@@ -1,3 +1,4 @@
+/* eslint-disable require-jsdoc */
 /* globals SaihuBot */
 'use strict';
 
@@ -11,13 +12,13 @@
  * @return {HTMLElement} response description with link
  */
 function renderResponse(action, term, url, engine) {
-  var span = document.createElement('span');
-  var line1 = document.createTextNode(`${action} "`);
-  let link = document.createElement('a');
+  const span = document.createElement('span');
+  const line1 = document.createTextNode(`${action} "`);
+  const link = document.createElement('a');
   link.href = url;
-  let linkText = document.createTextNode(`${term}`);
+  const linkText = document.createTextNode(`${term}`);
   link.appendChild(linkText);
-  var line2 = document.createTextNode(`" via ${engine}`);
+  const line2 = document.createTextNode(`" via ${engine}`);
   span.appendChild(line1);
   span.appendChild(link);
   span.appendChild(line2);
@@ -33,8 +34,8 @@ SaihuBot.prototype.responses.push({
   help: 'duckduckgo [term] - search [term] with DuckDuckgo',
   rule: /(^duckduckgo |^duck )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://duckduckgo.com/?q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'DuckDuckgo');
+    const url = 'https://duckduckgo.com/?q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'DuckDuckgo');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -45,8 +46,8 @@ SaihuBot.prototype.responses.push({
   help: 'wikipedia|w|wiki [term] - search wikipedia with [term]',
   rule: /(^wikipedia |^w |^wiki )(.*)/i,
   action: function(robot, msg) {
-    let url = 'http://en.wikipedia.org/w/index.php?title=Special:Search&search=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Wikipedia');
+    const url = 'http://en.wikipedia.org/w/index.php?title=Special:Search&search=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Wikipedia');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -57,8 +58,8 @@ SaihuBot.prototype.responses.push({
   help: 'github|gh [term] - search github with [term]',
   rule: /(^github |^gh )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://github.com/search?utf8=%E2%9C%93&q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Github');
+    const url = 'https://github.com/search?utf8=%E2%9C%93&q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Github');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -69,8 +70,8 @@ SaihuBot.prototype.responses.push({
   help: 'google|g [term] - google search with [term]',
   rule: /(^google |^g )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://www.google.com/search?q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Google');
+    const url = 'https://www.google.com/search?q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Google');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -81,8 +82,8 @@ SaihuBot.prototype.responses.push({
   help: 'translate|tr|trans|gt [term] - translate [term] with google translate',
   rule: /(^translate |^tr |^trans |^gt )(.*)/i,
   action: function(robot, msg) {
-    let url = 'http://translate.google.com/?text=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Translate', msg[2], url, 'Google Translate');
+    const url = 'http://translate.google.com/?text=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Translate', msg[2], url, 'Google Translate');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -93,8 +94,8 @@ SaihuBot.prototype.responses.push({
   help: 'images|image|gi [term] - search [term] with google images',
   rule: /(^images |^image |^gi )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://www.google.com/search?tbm=isch&q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Google Images');
+    const url = 'https://www.google.com/search?tbm=isch&q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Google Images');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -105,8 +106,8 @@ SaihuBot.prototype.responses.push({
   help: 'maps|map|gm [term] - search [term] with google maps',
   rule: /(^maps |^map |^gm )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://www.google.com/maps?q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Google Maps');
+    const url = 'https://www.google.com/maps?q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Google Maps');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -117,8 +118,8 @@ SaihuBot.prototype.responses.push({
   help: 'youtube|yt [term] - search [term] with Youtube',
   rule: /(^youtube |^yt )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://www.youtube.com/results?search_query=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Youtube');
+    const url = 'https://www.youtube.com/results?search_query=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Youtube');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -129,8 +130,8 @@ SaihuBot.prototype.responses.push({
   help: 'mdn [term] - search [term] with Mozilla Developer Network',
   rule: /(^mdn )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://mdn.io/' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'MDN');
+    const url = 'https://mdn.io/' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'MDN');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -141,8 +142,8 @@ SaihuBot.prototype.responses.push({
   help: 'npm [term] - search [term] with npm',
   rule: /(^npm )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://www.npmjs.com/search?q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'npm');
+    const url = 'https://www.npmjs.com/search?q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'npm');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -153,8 +154,8 @@ SaihuBot.prototype.responses.push({
   help: 'twitter [term] - search [term] with twitter',
   rule: /(^twitter )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://twitter.com/search?q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Twitter');
+    const url = 'https://twitter.com/search?q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Twitter');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -165,8 +166,8 @@ SaihuBot.prototype.responses.push({
   help: 'stackoverflow|stack|so [term] - search [term] with StackOverflow',
   rule: /(^stackoverflow |^stack |^so )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://stackoverflow.com/search?q=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'StackOverflow');
+    const url = 'https://stackoverflow.com/search?q=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'StackOverflow');
     robot.adapter.sendHTML(link);
     openTab(url);
   },
@@ -177,8 +178,8 @@ SaihuBot.prototype.responses.push({
   help: 'wolfram|wolf|wo [term] - search [term] with Wolfram Alpha',
   rule: /(^wolfram |^wolf |^wo |^siri )(.*)/i,
   action: function(robot, msg) {
-    let url = 'https://www.wolframalpha.com/input/?i=' + encodeURIComponent(msg[2]);
-    let link = renderResponse('Search', msg[2], url, 'Wolfram Alpha');
+    const url = 'https://www.wolframalpha.com/input/?i=' + encodeURIComponent(msg[2]);
+    const link = renderResponse('Search', msg[2], url, 'Wolfram Alpha');
     robot.adapter.sendHTML(link);
     openTab(url);
   },